// frontend/src/api/generated/model/normalizationMappingOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Family Finance Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { NormalizationMappingOutRawValue } from './normalizationMappingOutRawValue';
import type { NormalizationMappingOutAccountId } from './normalizationMappingOutAccountId';
import type { NormalizationMappingOutMerchant } from './normalizationMappingOutMerchant';

export interface NormalizationMappingOut {
  id: number;
  kind: string;
  raw_value?: NormalizationMappingOutRawValue;
  canonical_value: string;
  account_id: NormalizationMappingOutAccountId;
  merchant?: NormalizationMappingOutMerchant;
}